tests:Add test to check shim exits with failure on connectshim error.
The shim does not wait for responses for commands, but if it gets
as error reponse for the connectshim command from proxy,
then it should exit with failure exit code. The test passes a token
in base64url encoded format, different than the mock proxy expects,
causing the mock to send an error response to shim.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>