stdin: Handle binary data on stdin
We were using strlen on the stdin data, but this may be binary data.
Use the return value of the read instead for the size and include
the size of the payload as an extra parameter to send_proxy_message.
Change all calls to send_proxy_message to include the size
parameter we added.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>