signals: Change the json passed to what the proxy expects
The json expected by the proxy for signal handling changed in
the new implementation. Sync the changes as the signal handling
was implemented before the protocol was fully defined and
changes were not synced.

The payload api from the proxy should be probably be vendored
in the mock proxy, to be done in a separate commit.

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>