shim: Handle a SIGKILL signal
Because a SIGKILL cannot be caught from the targeted process, we
had to find a way to detect the end of this process.
This patch proposes a solution relying on a child that we spawn
on behalf of the parent process. The only purpose of that child
is to listen to a monitoring pipe that has been opened from the
parent. In case the child receives no bytes after he was waiting
on this pipe, this means the parent could not write to the pipe
before to exit, meaning we have detected a case of SIGKILL. Then,
the child forward this signal to the container process through the
proxy, and it exits immediately with a call to _exit() system call.
This allows the child to avoid calling into the parent atexit()
functions.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>